+++ async_v20/interface/order.py
@@ -87,6 +87,26 @@ class OrderInterface(object):
         """
         create an OrderRequest
 
+        Args:
+
+            trade_id: :class:`~async_v20.definitions.primitives.TradeID`
+            price: :class:`~async_v20.definitions.primitives.PriceValue`
+            type: :class:`~async_v20.definitions.primitives.OrderType`
+            client_trade_id: :class:`~async_v20.definitions.primitives.ClientID`
+            time_in_force: :class:`~async_v20.definitions.primitives.TimeInForce`
+            gtd_time: :class:`~async_v20.definitions.primitives.DateTime`
+            trigger_condition: :class:`~async_v20.definitions.primitives.OrderTriggerCondition`
+            client_extensions: :class:`~async_v20.definitions.types.ClientExtensions`
+            distance: :class:`~async_v20.definitions.primitives.PriceValue`
+            instrument: :class:`~async_v20.definitions.primitives.InstrumentName`
+            units: :class:`~async_v20.definitions.primitives.Unit`
+            price_bound: :class:`~async_v20.definitions.primitives.PriceValue`
+            position_fill: :class:`~async_v20.definitions.primitives.OrderPositionFill`
+            take_profit_on_fill: :class:`~async_v20.definitions.types.TakeProfitDetails`
+            stop_loss_on_fill: :class:`~async_v20.definitions.types.StopLossDetails`
+            trailing_stop_loss_on_fill: :class:`~async_v20.definitions.types.TrailingStopLossDetails`
+            trade_client_extensions: :class:`~async_v20.definitions.types.ClientExtensions`
+
         Returns:
 
             status [201]
@@ -138,7 +158,7 @@ class OrderInterface(object):
         Args:
 
             ids: :class:`~async_v20.endpoints.annotations.Ids`
-            list of Order IDs to retrieve
+                list of Order IDs to retrieve
             state: :class:`~async_v20.definitions.primitives.OrderStateFilter`
                 The state to filter the requested Orders by
             instrument: :class:`~async_v20.definitions.primitives.InstrumentName`
@@ -332,6 +352,44 @@ class OrderInterface(object):
         """
         Create a Market Order Request
 
+        Args:
+
+            instrument: :class:`~async_v20.definitions.primitives.InstrumentName`
+                The Market Order's Instrument.
+            units: :class:`~async_v20.definitions.primitives.Unit`
+                The quantity requested to be filled by the Market Order. A posititive number of units
+                results in a long Order, and a negative number of units results in a short Order.
+            time_in_force: :class:`~async_v20.definitions.primitives.TimeInForce`
+                The time-in-force requested for the Market Order.
+                Restricted to FOK or IOC for a MarketOrder.
+            price_bound: :class:`~async_v20.definitions.primitives.PriceValue`
+                The worst price that the client is willing to have the Market Order filled at.
+            position_fill: :class:`~async_v20.definitions.primitives.OrderPositionFill`
+                Specification of how Positions in the Account
+                are modified when the Order is filled.
+            client_extensions: :class:`~async_v20.definitions.types.ClientExtensions`
+                The client extensions to add to the Order. Do not set,
+                modify, or delete clientExtensions if your account is associated with MT4.
+            take_profit_on_fill: :class:`~async_v20.definitions.types.TakeProfitDetails`
+                TakeProfitDetails specifies the details of a Take Profit Order to be created on behalf of
+                a client. This may happen when an Order
+                is filled that opens a Trade requiring a Take Profit, or when a Trade's dependent Take Profit Order is
+                modified directly through the Trade.
+            stop_loss_on_fill: :class:`~async_v20.definitions.types.StopLossDetails`
+                StopLossDetails specifies the details of a Stop Loss Order to be created on behalf of a
+                client. This may happen when an Order
+                is filled that opens a Trade requiring a Stop Loss, or when a Trade's dependent Stop Loss Order is modified
+                directly through the Trade.
+            trailing_stop_loss_on_fill: :class:`~async_v20.definitions.types.TrailingStopLossDetails`
+                TrailingStopLossDetails specifies the details of a Trailing Stop Loss Order to be
+                created on behalf of a client. This may happen when an Order is
+                filled that opens a Trade requiring a Trailing Stop Loss, or when a Trade's dependent Trailing Stop Loss
+                Order is modified directly through the Trade.
+            trade_client_extensions: :class:`~async_v20.definitions.types.ClientExtensions`
+                Client Extensions to add to the Trade created when the Order is filled (if such a
+                Trade is created). Do not set, modify, or delete tradeClientExtensions if your account is associated with
+                MT4.
+
         Returns:
 
             status [201]
@@ -361,14 +419,15 @@ class OrderInterface(object):
                 errorMessage= :class:`~builtins.str`)
         """
         return self.post_order(
-            order_request=MarketOrderRequest(instrument=instrument, units=units, time_in_force=time_in_force,
-                                             price_bound=price_bound, position_fill=position_fill,
-                                             client_extensions=client_extensions,
-                                             take_profit_on_fill=take_profit_on_fill,
-                                             stop_loss_on_fill=stop_loss_on_fill,
-                                             trailing_stop_loss_on_fill=trailing_stop_loss_on_fill,
-                                             trade_client_extensions=trade_client_extensions
-                                             ))
+            order_request=MarketOrderRequest(
+                instrument=instrument, units=units, time_in_force=time_in_force,
+                price_bound=price_bound, position_fill=position_fill,
+                client_extensions=client_extensions,
+                take_profit_on_fill=take_profit_on_fill,
+                stop_loss_on_fill=stop_loss_on_fill,
+                trailing_stop_loss_on_fill=trailing_stop_loss_on_fill,
+                trade_client_extensions=trade_client_extensions
+            ))
 
     @shortcut
     def limit_order(self, instrument: InstrumentName, units: Unit, price: PriceValue,
@@ -380,6 +439,50 @@ class OrderInterface(object):
                     trade_client_extensions: ClientExtensions = ...):
         """
         Create a Limit Order
+
+        Args:
+
+            instrument: :class:`~async_v20.definitions.primitives.InstrumentName`
+                The Limit Order's Instrument.
+            units: :class:`~async_v20.definitions.primitives.Unit`
+                The quantity requested to be filled by the Limit Order. A posititive number of units
+                results in a long Order, and a negative number of units results in a short Order.
+            price: :class:`~async_v20.definitions.primitives.PriceValue`
+                The price threshold specified for the Limit Order. The Limit Order will only be
+                filled by a market price that is equal to or better than this price.
+            time_in_force: :class:`~async_v20.definitions.primitives.TimeInForce`
+                The time-in-force requested for the Limit Order.
+            gtd_time: :class:`~async_v20.definitions.primitives.DateTime`
+                The date/time when the Limit Order will
+                be cancelled if its timeInForce is "GTD".
+            position_fill: :class:`~async_v20.definitions.primitives.OrderPositionFill`
+                Specification of how Positions in the Account
+                are modified when the Order is filled.
+            trigger_condition: :class:`~async_v20.definitions.primitives.OrderTriggerCondition`
+                Specification of what component of a price should be used
+                for comparison when determining if the Order should be filled.
+            client_extensions: :class:`~async_v20.definitions.types.ClientExtensions`
+                The client extensions to add to the Order. Do not set,
+                modify, or delete clientExtensions if your account is associated with MT4.
+            take_profit_on_fill: :class:`~async_v20.definitions.types.TakeProfitDetails`
+                TakeProfitDetails specifies the details of a Take Profit Order to be created on behalf of
+                a client. This may happen when an Order
+                is filled that opens a Trade requiring a Take Profit, or when a Trade's dependent Take Profit Order is
+                modified directly through the Trade.
+            stop_loss_on_fill: :class:`~async_v20.definitions.types.StopLossDetails`
+                StopLossDetails specifies the details of a Stop Loss Order to be created on behalf of a
+                client. This may happen when an Order
+                is filled that opens a Trade requiring a Stop Loss, or when a Trade's dependent Stop Loss Order is modified
+                directly through the Trade.
+            trailing_stop_loss_on_fill: :class:`~async_v20.definitions.types.TrailingStopLossDetails`
+                TrailingStopLossDetails specifies the details of a Trailing Stop Loss Order to be
+                created on behalf of a client. This may happen when an Order is
+                filled that opens a Trade requiring a Trailing Stop Loss, or when a Trade's dependent Trailing Stop Loss
+                Order is modified directly through the Trade.
+            trade_client_extensions: :class:`~async_v20.definitions.types.ClientExtensions`
+                Client Extensions to add to the Trade created when the Order is filled (if such a
+                Trade is created). Do not set, modify, or delete tradeClientExtensions if your account is associated with
+                MT4.
 
         Returns:
 
@@ -409,30 +512,76 @@ class OrderInterface(object):
                 errorCode= :class:`~builtins.str`,
                 errorMessage= :class:`~builtins.str`)
         """
-        return self.post_order(order_request=LimitOrderRequest(
-            instrument=instrument, units=units, price=price,
-            time_in_force=time_in_force, gtd_time=gtd_time,
-            position_fill=position_fill,
-            trigger_condition=trigger_condition,
-            client_extensions=client_extensions,
-            take_profit_on_fill=take_profit_on_fill,
-            stop_loss_on_fill=stop_loss_on_fill,
-            trailing_stop_loss_on_fill=trailing_stop_loss_on_fill,
-            trade_client_extensions=trade_client_extensions
-        ))
+        return self.post_order(
+            order_request=LimitOrderRequest(
+                instrument=instrument, units=units, price=price,
+                time_in_force=time_in_force, gtd_time=gtd_time,
+                position_fill=position_fill,
+                trigger_condition=trigger_condition,
+                client_extensions=client_extensions,
+                take_profit_on_fill=take_profit_on_fill,
+                stop_loss_on_fill=stop_loss_on_fill,
+                trailing_stop_loss_on_fill=trailing_stop_loss_on_fill,
+                trade_client_extensions=trade_client_extensions
+            ))
 
-    @endpoint(PUTOrderSpecifier)
+    @shortcut
     def limit_replace_order(self,
-                            order_specifier: OrderSpecifier,
-                            order_request: LimitOrderRequest):
+                            order_specifier: OrderSpecifier, instrument: InstrumentName, units: Unit, price: PriceValue,
+                            time_in_force: TimeInForce = 'GTC', gtd_time: DateTime = ...,
+                            position_fill: OrderPositionFill = 'DEFAULT',
+                            trigger_condition: OrderTriggerCondition = 'DEFAULT',
+                            client_extensions: ClientExtensions = ..., take_profit_on_fill: TakeProfitDetails = ...,
+                            stop_loss_on_fill: StopLossDetails = ...,
+                            trailing_stop_loss_on_fill: TrailingStopLossDetails = ...,
+                            trade_client_extensions: ClientExtensions = ...):
         """
         Replace a pending Limit Order
 
         Args:
             order_specifier: :class:`~async_v20.definitions.primitives.OrderSpecifier`
                 The ID of the Limit Order to replace
-            order: :class:`~async_v20.definitions.types.LimitOrderRequest`
-                The arguments to create a LimitOrderRequest
+            instrument: :class:`~async_v20.definitions.primitives.InstrumentName`
+                The Limit Order's Instrument.
+            units: :class:`~async_v20.definitions.primitives.Unit`
+                The quantity requested to be filled by the Limit Order. A posititive number of units
+                results in a long Order, and a negative number of units results in a short Order.
+            price: :class:`~async_v20.definitions.primitives.PriceValue`
+                The price threshold specified for the Limit Order. The Limit Order will only be
+                filled by a market price that is equal to or better than this price.
+            time_in_force: :class:`~async_v20.definitions.primitives.TimeInForce`
+                The time-in-force requested for the Limit Order.
+            gtd_time: :class:`~async_v20.definitions.primitives.DateTime`
+                The date/time when the Limit Order will
+                be cancelled if its timeInForce is "GTD".
+            position_fill: :class:`~async_v20.definitions.primitives.OrderPositionFill`
+                Specification of how Positions in the Account
+                are modified when the Order is filled.
+            trigger_condition: :class:`~async_v20.definitions.primitives.OrderTriggerCondition`
+                Specification of what component of a price should be used
+                for comparison when determining if the Order should be filled.
+            client_extensions: :class:`~async_v20.definitions.types.ClientExtensions`
+                The client extensions to add to the Order. Do not set,
+                modify, or delete clientExtensions if your account is associated with MT4.
+            take_profit_on_fill: :class:`~async_v20.definitions.types.TakeProfitDetails`
+                TakeProfitDetails specifies the details of a Take Profit Order to be created on behalf of
+                a client. This may happen when an Order
+                is filled that opens a Trade requiring a Take Profit, or when a Trade's dependent Take Profit Order is
+                modified directly through the Trade.
+            stop_loss_on_fill: :class:`~async_v20.definitions.types.StopLossDetails`
+                StopLossDetails specifies the details of a Stop Loss Order to be created on behalf of a
+                client. This may happen when an Order
+                is filled that opens a Trade requiring a Stop Loss, or when a Trade's dependent Stop Loss Order is modified
+                directly through the Trade.
+            trailing_stop_loss_on_fill: :class:`~async_v20.definitions.types.TrailingStopLossDetails`
+                TrailingStopLossDetails specifies the details of a Trailing Stop Loss Order to be
+                created on behalf of a client. This may happen when an Order is
+                filled that opens a Trade requiring a Trailing Stop Loss, or when a Trade's dependent Trailing Stop Loss
+                Order is modified directly through the Trade.
+            trade_client_extensions: :class:`~async_v20.definitions.types.ClientExtensions`
+                Client Extensions to add to the Trade created when the Order is filled (if such a
+                Trade is created). Do not set, modify, or delete tradeClientExtensions if your account is associated with
+                MT4.
 
         Returns:
 
@@ -463,7 +612,17 @@ class OrderInterface(object):
                 errorCode= :class:`~builtins.str`,
                 errorMessage= :class:`~builtins.str`)
         """
-        pass
+        return self.replace_order(
+            order_specifier=order_specifier,
+            order_request=LimitOrderRequest(
+                instrument=instrument, units=units, price=price,
+                time_in_force=time_in_force, gtd_time=gtd_time, position_fill=position_fill,
+                trigger_condition=trigger_condition, client_extensions=client_extensions,
+                take_profit_on_fill=take_profit_on_fill,
+                stop_loss_on_fill=stop_loss_on_fill,
+                trailing_stop_loss_on_fill=trailing_stop_loss_on_fill,
+                trade_client_extensions=trade_client_extensions
+            ))
 
     @shortcut
     def stop_order(self, trade_id: TradeID, price: PriceValue,
@@ -471,6 +630,28 @@ class OrderInterface(object):
                    trigger_condition: OrderTriggerCondition = 'DEFAULT', client_extensions: ClientExtensions = ...):
         """
         Create a Stop Order
+
+        Args:
+
+            trade_id: :class:`~async_v20.definitions.primitives.TradeID`
+                The ID of the Trade to close when the price threshold is breached.
+            client_trade_id: :class:`~async_v20.definitions.primitives.TradeID`
+                The client ID of the Trade to be closed when the price threshold is breached.
+            price: :class:`~async_v20.definitions.primitives.PriceValue`
+                The price threshold specified for the StopLoss Order. The associated Trade will be
+                closed by a market price that is equal to or worse than this threshold.
+            time_in_force: :class:`~async_v20.definitions.primitives.TimeInForce`
+                The time-in-force requested for the StopLoss Order. Restricted
+                to "GTC", "GFD" and "GTD" for StopLoss Orders.
+            gtd_time: :class:`~async_v20.definitions.primitives.DateTime`
+                The date/time when the StopLoss Order will
+                be cancelled if its timeInForce is "GTD".
+            trigger_condition: :class:`~async_v20.definitions.primitives.OrderTriggerCondition`
+                Specification of what component of a price should be used
+                for comparison when determining if the Order should be filled.
+            client_extensions: :class:`~async_v20.definitions.types.ClientExtensions`
+                The client extensions to add to the Order. Do not set,
+                modify, or delete clientExtensions if your account is associated with MT4.
 
         Returns:
 
@@ -522,10 +703,53 @@ class OrderInterface(object):
         Replace a pending Stop Order
 
         Args:
+
             order_specifier: :class:`~async_v20.definitions.primitives.OrderSpecifier`
                 The ID of the Stop Order to replace
-            order: :class:`~async_v20.definitions.types.StopOrderRequest`
-                The arguments to create a StopOrderRequest
+            instrument: :class:`~async_v20.definitions.primitives.InstrumentName`
+                The Stop Order's Instrument.
+            units: :class:`~async_v20.definitions.primitives.Unit`
+                The quantity requested to be filled by the Stop Order. A posititive number of units
+                results in a long Order, and a negative number of units results in a short Order.
+            price: :class:`~async_v20.definitions.primitives.PriceValue`
+                The price threshold specified for the Stop Order. The Stop Order will only be
+                filled by a market price that is equal to or worse than this price.
+            price_bound: :class:`~async_v20.definitions.primitives.PriceValue`
+                The worst market price that may be used to fill this Stop Order. If the market gaps and
+                crosses through both the price and the priceBound, the Stop Order will be cancelled instead of being filled.
+            time_in_force: :class:`~async_v20.definitions.primitives.TimeInForce`
+                The time-in-force requested for the Stop Order.
+            gtd_time: :class:`~async_v20.definitions.primitives.DateTime`
+                The date/time when the Stop Order will
+                be cancelled if its timeInForce is "GTD".
+            position_fill: :class:`~async_v20.definitions.primitives.OrderPositionFill`
+                Specification of how Positions in the Account
+                are modified when the Order is filled.
+            trigger_condition: :class:`~async_v20.definitions.primitives.OrderTriggerCondition`
+                Specification of what component of a price should be used
+                for comparison when determining if the Order should be filled.
+            client_extensions: :class:`~async_v20.definitions.types.ClientExtensions`
+                The client extensions to add to the Order. Do not set,
+                modify, or delete clientExtensions if your account is associated with MT4.
+            take_profit_on_fill: :class:`~async_v20.definitions.types.TakeProfitDetails`
+                TakeProfitDetails specifies the details of a Take Profit Order to be created on behalf of
+                a client. This may happen when an Order
+                is filled that opens a Trade requiring a Take Profit, or when a Trade's dependent Take Profit Order is
+                modified directly through the Trade.
+            stop_loss_on_fill: :class:`~async_v20.definitions.types.StopLossDetails`
+                StopLossDetails specifies the details of a Stop Loss Order to be created on behalf of a
+                client. This may happen when an Order
+                is filled that opens a Trade requiring a Stop Loss, or when a Trade's dependent Stop Loss Order is modified
+                directly through the Trade.
+            trailing_stop_loss_on_fill: :class:`~async_v20.definitions.types.TrailingStopLossDetails`
+                TrailingStopLossDetails specifies the details of a Trailing Stop Loss Order to be
+                created on behalf of a client. This may happen when an Order is
+                filled that opens a Trade requiring a Trailing Stop Loss, or when a Trade's dependent Trailing Stop Loss
+                Order is modified directly through the Trade.
+            trade_client_extensions: :class:`~async_v20.definitions.types.ClientExtensions`
+                Client Extensions to add to the Trade created when the Order is filled (if such a
+                Trade is created). Do not set, modify, or delete tradeClientExtensions if your account is associated with
+                MT4.
 
         Returns:
 
@@ -557,18 +781,19 @@ class OrderInterface(object):
                 errorMessage= :class:`~builtins.str`)
 
         """
-        return self.replace_order(order_specifier=order_specifier,
-                           order_request=StopOrderRequest(
-                               instrument=instrument, units=units, price=price,
-                               price_bound=price_bound, time_in_force=time_in_force,
-                               gtd_time=gtd_time, position_fill=position_fill,
-                               trigger_condition=trigger_condition,
-                               client_extensions=client_extensions,
-                               take_profit_on_fill=take_profit_on_fill,
-                               stop_loss_on_fill=stop_loss_on_fill,
-                               trailing_stop_loss_on_fill=trailing_stop_loss_on_fill,
-                               trade_client_extensions=trade_client_extensions
-                           ))
+        return self.replace_order(
+            order_specifier=order_specifier,
+            order_request=StopOrderRequest(
+                instrument=instrument, units=units, price=price,
+                price_bound=price_bound, time_in_force=time_in_force,
+                gtd_time=gtd_time, position_fill=position_fill,
+                trigger_condition=trigger_condition,
+                client_extensions=client_extensions,
+                take_profit_on_fill=take_profit_on_fill,
+                stop_loss_on_fill=stop_loss_on_fill,
+                trailing_stop_loss_on_fill=trailing_stop_loss_on_fill,
+                trade_client_extensions=trade_client_extensions
+            ))
 
     @shortcut
     def market_if_touched_order(self, instrument: InstrumentName, units: Unit, price: PriceValue,
@@ -583,6 +808,56 @@ class OrderInterface(object):
                                 trade_client_extensions: ClientExtensions = ...):
         """
         Create a market if touched order
+
+        Args:
+
+            instrument: :class:`~async_v20.definitions.primitives.InstrumentName`
+                The MarketIfTouched Order's Instrument.
+            units: :class:`~async_v20.definitions.primitives.Unit`
+                The quantity requested to be filled by the MarketIfTouched Order. A posititive number of units
+                results in a long Order, and a negative number of units results in a short Order.
+            price: :class:`~async_v20.definitions.primitives.PriceValue`
+                The price threshold specified for the MarketIfTouched Order. The MarketIfTouched Order will only be
+                filled by a market price that crosses this price from the direction of the market price
+                at the time when the Order was created (the initialMarketPrice). Depending on the value of the Order's
+                price and initialMarketPrice, the MarketIfTouchedOrder will behave like a Limit or a Stop Order.
+            price_bound: :class:`~async_v20.definitions.primitives.PriceValue`
+                The worst market price that may be used to fill this MarketIfTouched Order.
+            time_in_force: :class:`~async_v20.definitions.primitives.TimeInForce`
+                The time-in-force requested for the MarketIfTouched Order. Restricted
+                to "GTC", "GFD" and "GTD" for MarketIfTouched Orders.
+            gtd_time: :class:`~async_v20.definitions.primitives.DateTime`
+                The date/time when the MarketIfTouched Order will
+                be cancelled if its timeInForce is "GTD".
+            position_fill: :class:`~async_v20.definitions.primitives.OrderPositionFill`
+                Specification of how Positions in the Account
+                are modified when the Order is filled.
+            trigger_condition: :class:`~async_v20.definitions.primitives.OrderTriggerCondition`
+                Specification of what component of a price should be used
+                for comparison when determining if the Order should be filled.
+            client_extensions: :class:`~async_v20.definitions.types.ClientExtensions`
+                The client extensions to add to the Order. Do not set,
+                modify, or delete clientExtensions if your account is associated with MT4.
+            take_profit_on_fill: :class:`~async_v20.definitions.types.TakeProfitDetails`
+                TakeProfitDetails specifies the details of a Take Profit Order to be created on behalf of
+                a client. This may happen when an Order
+                is filled that opens a Trade requiring a Take Profit, or when a Trade's dependent Take Profit Order is
+                modified directly through the Trade.
+            stop_loss_on_fill: :class:`~async_v20.definitions.types.StopLossDetails`
+                StopLossDetails specifies the details of a Stop Loss Order to be created on behalf of a
+                client. This may happen when an Order
+                is filled that opens a Trade requiring a Stop Loss, or when a Trade's dependent Stop Loss Order is modified
+                directly through the Trade.
+            trailing_stop_loss_on_fill: :class:`~async_v20.definitions.types.TrailingStopLossDetails`
+                TrailingStopLossDetails specifies the details of a Trailing Stop Loss Order to be
+                created on behalf of a client. This may happen when an Order is
+                filled that opens a Trade requiring a Trailing Stop Loss, or when a Trade's dependent Trailing Stop Loss
+                Order is modified directly through the Trade.
+            trade_client_extensions: :class:`~async_v20.definitions.types.ClientExtensions`
+                Client Extensions to add to the Trade created when the Order is filled (if such a
+                Trade is created). Do not set, modify, or delete tradeClientExtensions if your account is associated with
+                MT4.
+
 
         Returns:
 
@@ -643,10 +918,55 @@ class OrderInterface(object):
         Replace a pending market if touched order
 
         Args:
+
             order_specifier: :class:`~async_v20.definitions.primitives.OrderSpecifier`
                 The ID of the MarketIfTouched Order to replace
-            order: :class:`~async_v20.definitions.types.MarketIfTouchedOrderRequest`
-                The arguments to create a MarketIfTouchedOrderRequest
+            instrument: :class:`~async_v20.definitions.primitives.InstrumentName`
+                The MarketIfTouched Order's Instrument.
+            units: :class:`~async_v20.definitions.primitives.Unit`
+                The quantity requested to be filled by the MarketIfTouched Order. A posititive number of units
+                results in a long Order, and a negative number of units results in a short Order.
+            price: :class:`~async_v20.definitions.primitives.PriceValue`
+                The price threshold specified for the MarketIfTouched Order. The MarketIfTouched Order will only be
+                filled by a market price that crosses this price from the direction of the market price
+                at the time when the Order was created (the initialMarketPrice). Depending on the value of the Order's
+                price and initialMarketPrice, the MarketIfTouchedOrder will behave like a Limit or a Stop Order.
+            price_bound: :class:`~async_v20.definitions.primitives.PriceValue`
+                The worst market price that may be used to fill this MarketIfTouched Order.
+            time_in_force: :class:`~async_v20.definitions.primitives.TimeInForce`
+                The time-in-force requested for the MarketIfTouched Order. Restricted
+                to "GTC", "GFD" and "GTD" for MarketIfTouched Orders.
+            gtd_time: :class:`~async_v20.definitions.primitives.DateTime`
+                The date/time when the MarketIfTouched Order will
+                be cancelled if its timeInForce is "GTD".
+            position_fill: :class:`~async_v20.definitions.primitives.OrderPositionFill`
+                Specification of how Positions in the Account
+                are modified when the Order is filled.
+            trigger_condition: :class:`~async_v20.definitions.primitives.OrderTriggerCondition`
+                Specification of what component of a price should be used
+                for comparison when determining if the Order should be filled.
+            client_extensions: :class:`~async_v20.definitions.types.ClientExtensions`
+                The client extensions to add to the Order. Do not set,
+                modify, or delete clientExtensions if your account is associated with MT4.
+            take_profit_on_fill: :class:`~async_v20.definitions.types.TakeProfitDetails`
+                TakeProfitDetails specifies the details of a Take Profit Order to be created on behalf of
+                a client. This may happen when an Order
+                is filled that opens a Trade requiring a Take Profit, or when a Trade's dependent Take Profit Order is
+                modified directly through the Trade.
+            stop_loss_on_fill: :class:`~async_v20.definitions.types.StopLossDetails`
+                StopLossDetails specifies the details of a Stop Loss Order to be created on behalf of a
+                client. This may happen when an Order
+                is filled that opens a Trade requiring a Stop Loss, or when a Trade's dependent Stop Loss Order is modified
+                directly through the Trade.
+            trailing_stop_loss_on_fill: :class:`~async_v20.definitions.types.TrailingStopLossDetails`
+                TrailingStopLossDetails specifies the details of a Trailing Stop Loss Order to be
+                created on behalf of a client. This may happen when an Order is
+                filled that opens a Trade requiring a Trailing Stop Loss, or when a Trade's dependent Trailing Stop Loss
+                Order is modified directly through the Trade.
+            trade_client_extensions: :class:`~async_v20.definitions.types.ClientExtensions`
+                Client Extensions to add to the Trade created when the Order is filled (if such a
+                Trade is created). Do not set, modify, or delete tradeClientExtensions if your account is associated with
+                MT4.
 
         Returns:
 
@@ -678,20 +998,21 @@ class OrderInterface(object):
                 errorMessage= :class:`~builtins.str`)
 
         """
-        return self.replace_order(order_specifier=order_specifier,
-                                  order_request=MarketIfTouchedOrderRequest(
-                                      instrument=instrument, units=units,
-                                      price=price, price_bound=price_bound,
-                                      time_in_force=time_in_force,
-                                      gtd_time=gtd_time,
-                                      position_fill=position_fill,
-                                      trigger_condition=trigger_condition,
-                                      client_extensions=client_extensions,
-                                      take_profit_on_fill=take_profit_on_fill,
-                                      stop_loss_on_fill=stop_loss_on_fill,
-                                      trailing_stop_loss_on_fill=trailing_stop_loss_on_fill,
-                                      trade_client_extensions=trade_client_extensions)
-                                  )
+        return self.replace_order(
+            order_specifier=order_specifier,
+            order_request=MarketIfTouchedOrderRequest(
+                instrument=instrument, units=units,
+                price=price, price_bound=price_bound,
+                time_in_force=time_in_force,
+                gtd_time=gtd_time,
+                position_fill=position_fill,
+                trigger_condition=trigger_condition,
+                client_extensions=client_extensions,
+                take_profit_on_fill=take_profit_on_fill,
+                stop_loss_on_fill=stop_loss_on_fill,
+                trailing_stop_loss_on_fill=trailing_stop_loss_on_fill,
+                trade_client_extensions=trade_client_extensions)
+        )
 
     @shortcut
     def take_profit_order(self, trade_id: TradeID, price: PriceValue,
@@ -701,6 +1022,28 @@ class OrderInterface(object):
                           client_extensions: ClientExtensions = ...):
         """
         Create a take profit order
+
+        Args:
+
+            trade_id: :class:`~async_v20.definitions.primitives.TradeID`
+                The ID of the Trade to close when the price threshold is breached.
+            client_trade_id: :class:`~async_v20.definitions.primitives.TradeID`
+                The client ID of the Trade to be closed when the price threshold is breached.
+            price: :class:`~async_v20.definitions.primitives.PriceValue`
+                The price threshold specified for the TakeProfit Order. The associated Trade will be
+                closed by a market price that is equal to or better than this threshold.
+            time_in_force: :class:`~async_v20.definitions.primitives.TimeInForce`
+                The time-in-force requested for the TakeProfit Order. Restricted
+                to "GTC", "GFD" and "GTD" for TakeProfit Orders.
+            gtd_time: :class:`~async_v20.definitions.primitives.DateTime`
+                The date/time when the TakeProfit Order will
+                be cancelled if its timeInForce is "GTD".
+            trigger_condition: :class:`~async_v20.definitions.primitives.OrderTriggerCondition`
+                Specification of what component of a price should be used
+                for comparison when determining if the Order should be filled.
+            client_extensions: :class:`~async_v20.definitions.types.ClientExtensions`
+                The client extensions to add to the Order. Do not set,
+                modify, or delete clientExtensions if your account is associated with MT4.
 
         Returns:
 
@@ -755,7 +1098,26 @@ class OrderInterface(object):
 
             order_specifier: :class:`~async_v20.definitions.primitives.OrderSpecifier`
                 The ID of the Take Profit Order to replace
-            order: :class:`~async_v20.definitions.types.TakeProfitOrderRequest`
+            trade_id: :class:`~async_v20.definitions.primitives.TradeID`
+                The ID of the Trade to close when the price threshold is breached.
+            client_trade_id: :class:`~async_v20.definitions.primitives.TradeID`
+                The client ID of the Trade to be closed when the price threshold is breached.
+            price: :class:`~async_v20.definitions.primitives.PriceValue`
+                The price threshold specified for the TakeProfit Order. The associated Trade will be
+                closed by a market price that is equal to or better than this threshold.
+            time_in_force: :class:`~async_v20.definitions.primitives.TimeInForce`
+                The time-in-force requested for the TakeProfit Order. Restricted
+                to "GTC", "GFD" and "GTD" for TakeProfit Orders.
+            gtd_time: :class:`~async_v20.definitions.primitives.DateTime`
+                The date/time when the TakeProfit Order will
+                be cancelled if its timeInForce is "GTD".
+            trigger_condition: :class:`~async_v20.definitions.primitives.OrderTriggerCondition`
+                Specification of what component of a price should be used
+                for comparison when determining if the Order should be filled.
+            client_extensions: :class:`~async_v20.definitions.types.ClientExtensions`
+                The client extensions to add to the Order. Do not set,
+                modify, or delete clientExtensions if your account is associated with MT4.
+
 
         Returns:
 
@@ -786,14 +1148,15 @@ class OrderInterface(object):
                 errorCode= :class:`~builtins.str`,
                 errorMessage= :class:`~builtins.str`)
         """
-        return self.replace_order(order_specifier=order_specifier,
-                                  order_request=TakeProfitOrderRequest(
-                                      trade_id=trade_id, price=price,
-                                      client_trade_id=client_trade_id,
-                                      time_in_force=time_in_force, gtd_time=gtd_time,
-                                      trigger_condition=trigger_condition,
-                                      client_extensions=client_extensions)
-                                  )
+        return self.replace_order(
+            order_specifier=order_specifier,
+            order_request=TakeProfitOrderRequest(
+                trade_id=trade_id, price=price,
+                client_trade_id=client_trade_id,
+                time_in_force=time_in_force, gtd_time=gtd_time,
+                trigger_condition=trigger_condition,
+                client_extensions=client_extensions)
+        )
 
     @shortcut
     def stop_loss_order(self, trade_id: TradeID, price: PriceValue,
@@ -802,6 +1165,28 @@ class OrderInterface(object):
                         client_extensions: ClientExtensions = ...):
         """
         Create a Stop Loss Order
+
+        Args:
+
+            trade_id: :class:`~async_v20.definitions.primitives.TradeID`
+                The ID of the Trade to close when the price threshold is breached.
+            client_trade_id: :class:`~async_v20.definitions.primitives.TradeID`
+                The client ID of the Trade to be closed when the price threshold is breached.
+            price: :class:`~async_v20.definitions.primitives.PriceValue`
+                The price threshold specified for the StopLoss Order. The associated Trade will be
+                closed by a market price that is equal to or worse than this threshold.
+            time_in_force: :class:`~async_v20.definitions.primitives.TimeInForce`
+                The time-in-force requested for the StopLoss Order. Restricted
+                to "GTC", "GFD" and "GTD" for StopLoss Orders.
+            gtd_time: :class:`~async_v20.definitions.primitives.DateTime`
+                The date/time when the StopLoss Order will
+                be cancelled if its timeInForce is "GTD".
+            trigger_condition: :class:`~async_v20.definitions.primitives.OrderTriggerCondition`
+                Specification of what component of a price should be used
+                for comparison when determining if the Order should be filled.
+            client_extensions: :class:`~async_v20.definitions.types.ClientExtensions`
+                The client extensions to add to the Order. Do not set,
+                modify, or delete clientExtensions if your account is associated with MT4.
 
         Returns:
 
@@ -852,7 +1237,25 @@ class OrderInterface(object):
 
             order_specifier: :class:`~async_v20.definitions.primitives.OrderSpecifier`
                 The ID of the Stop Loss Order to replace
-            order: :class:`~async_v20.definitions.types.StopLossOrderRequest`
+            trade_id: :class:`~async_v20.definitions.primitives.TradeID`
+                The ID of the Trade to close when the price threshold is breached.
+            client_trade_id: :class:`~async_v20.definitions.primitives.TradeID`
+                The client ID of the Trade to be closed when the price threshold is breached.
+            price: :class:`~async_v20.definitions.primitives.PriceValue`
+                The price threshold specified for the StopLoss Order. The associated Trade will be
+                closed by a market price that is equal to or worse than this threshold.
+            time_in_force: :class:`~async_v20.definitions.primitives.TimeInForce`
+                The time-in-force requested for the StopLoss Order. Restricted
+                to "GTC", "GFD" and "GTD" for StopLoss Orders.
+            gtd_time: :class:`~async_v20.definitions.primitives.DateTime`
+                The date/time when the StopLoss Order will
+                be cancelled if its timeInForce is "GTD".
+            trigger_condition: :class:`~async_v20.definitions.primitives.OrderTriggerCondition`
+                Specification of what component of a price should be used
+                for comparison when determining if the Order should be filled.
+            client_extensions: :class:`~async_v20.definitions.types.ClientExtensions`
+                The client extensions to add to the Order. Do not set,
+                modify, or delete clientExtensions if your account is associated with MT4.
 
         Returns:
 
@@ -884,14 +1287,15 @@ class OrderInterface(object):
                 errorMessage= :class:`~builtins.str`)
 
         """
-        return self.replace_order(order_specifier=order_specifier,
-                                  order_request=StopLossOrderRequest(
-                                      trade_id=trade_id, price=price,
-                                      client_trade_id=client_trade_id,
-                                      time_in_force=time_in_force, gtd_time=gtd_time,
-                                      trigger_condition=trigger_condition,
-                                      client_extensions=client_extensions
-                                  ))
+        return self.replace_order(
+            order_specifier=order_specifier,
+            order_request=StopLossOrderRequest(
+                trade_id=trade_id, price=price,
+                client_trade_id=client_trade_id,
+                time_in_force=time_in_force, gtd_time=gtd_time,
+                trigger_condition=trigger_condition,
+                client_extensions=client_extensions
+            ))
 
     @shortcut
     def trailing_stop_loss_order(self, trade_id: TradeID, distance: PriceValue,
@@ -901,6 +1305,28 @@ class OrderInterface(object):
                                  client_extensions: ClientExtensions = ...):
         """
         Create a Trailing Stop Loss Order
+
+        Args:
+
+            trade_id: :class:`~async_v20.definitions.primitives.TradeID`
+                The ID of the Trade to close when the price threshold is breached.
+            client_trade_id: :class:`~async_v20.definitions.primitives.TradeID`
+                The client ID of the Trade to be closed when the price threshold is breached.
+            distance: :class:`~async_v20.definitions.primitives.PriceValue`
+                The price distance specified for the TrailingStopLoss Order.
+            time_in_force: :class:`~async_v20.definitions.primitives.TimeInForce`
+                The time-in-force requested for the TrailingStopLoss Order. Restricted
+                to "GTC", "GFD" and "GTD" for TrailingStopLoss Orders.
+            gtd_time: :class:`~async_v20.definitions.primitives.DateTime`
+                The date/time when the StopLoss Order will
+                be cancelled if its timeInForce is "GTD".
+            trigger_condition: :class:`~async_v20.definitions.primitives.OrderTriggerCondition`
+                Specification of what component of a price should be used
+                for comparison when determining if the Order should be filled.
+            client_extensions: :class:`~async_v20.definitions.types.ClientExtensions`
+                The client extensions to add to the Order. Do not set,
+                modify, or delete clientExtensions if your account is associated with MT4.
+
 
         Returns:
 
@@ -930,14 +1356,15 @@ class OrderInterface(object):
                 errorCode= :class:`~builtins.str`,
                 errorMessage= :class:`~builtins.str`)
         """
-        return self.post_order(order_request=TrailingStopLossOrderRequest(
-            trade_id=trade_id, distance=distance,
-            client_trade_id=client_trade_id,
-            time_in_force=time_in_force,
-            gtd_time=gtd_time,
-            trigger_condition=trigger_condition,
-            client_extensions=client_extensions
-        ))
+        return self.post_order(
+            order_request=TrailingStopLossOrderRequest(
+                trade_id=trade_id, distance=distance,
+                client_trade_id=client_trade_id,
+                time_in_force=time_in_force,
+                gtd_time=gtd_time,
+                trigger_condition=trigger_condition,
+                client_extensions=client_extensions
+            ))
 
     @shortcut
     def trailing_stop_loss_replace_order(self, order_specifier: OrderSpecifier,
@@ -953,7 +1380,25 @@ class OrderInterface(object):
 
             order_specifier: :class:`~async_v20.definitions.primitives.OrderSpecifier`
                 The ID of the Take Profit Order to replace
-            order: :class:`~async_v20.definitions.types.TrailingStopLossOrderRequest`
+            trade_id: :class:`~async_v20.definitions.primitives.TradeID`
+                The ID of the Trade to close when the price threshold is breached.
+            client_trade_id: :class:`~async_v20.definitions.primitives.TradeID`
+                The client ID of the Trade to be closed when the price threshold is breached.
+            distance: :class:`~async_v20.definitions.primitives.PriceValue`
+                The price distance specified for the TrailingStopLoss Order.
+            time_in_force: :class:`~async_v20.definitions.primitives.TimeInForce`
+                The time-in-force requested for the TrailingStopLoss Order. Restricted
+                to "GTC", "GFD" and "GTD" for TrailingStopLoss Orders.
+            gtd_time: :class:`~async_v20.definitions.primitives.DateTime`
+                The date/time when the StopLoss Order will
+                be cancelled if its timeInForce is "GTD".
+            trigger_condition: :class:`~async_v20.definitions.primitives.OrderTriggerCondition`
+                Specification of what component of a price should be used
+                for comparison when determining if the Order should be filled.
+            client_extensions: :class:`~async_v20.definitions.types.ClientExtensions`
+                The client extensions to add to the Order. Do not set,
+                modify, or delete clientExtensions if your account is associated with MT4.
+
 
         Returns:
 
@@ -984,12 +1429,13 @@ class OrderInterface(object):
                 errorCode= :class:`~builtins.str`,
                 errorMessage= :class:`~builtins.str`)
         """
-        return self.replace_order(order_specifier=order_specifier,
-                                  order_request=TrailingStopLossOrderRequest(
-                                      trade_id=trade_id, distance=distance,
-                                      client_trade_id=client_trade_id,
-                                      time_in_force=time_in_force,
-                                      gtd_time=gtd_time,
-                                      trigger_condition=trigger_condition,
-                                      client_extensions=client_extensions
-                                  ))
+        return self.replace_order(
+            order_specifier=order_specifier,
+            order_request=TrailingStopLossOrderRequest(
+                trade_id=trade_id, distance=distance,
+                client_trade_id=client_trade_id,
+                time_in_force=time_in_force,
+                gtd_time=gtd_time,
+                trigger_condition=trigger_condition,
+                client_extensions=client_extensions
+            ))